Airflow/requirements.txt
apache-airflow==3.0.1
apache-airflow-providers-postgres==6.1.3
apache-airflow-providers-mongo==5.1.0
airflow-providers-clickhouse==0.0.1
clickhouse-driver==0.2.7
clickhouse-sqlalchemy==0.2.7
psycopg2-binary==2.9.9
pymongo==4.6.2
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
aiohttp==3.9.3
